Move Switch in its file
