+++ roseau/load_flow/models/__init__.py
@@ -9,7 +9,7 @@ from roseau.load_flow.models.branches import AbstractBranch
 from roseau.load_flow.models.buses import Bus
 from roseau.load_flow.models.core import Element
 from roseau.load_flow.models.grounds import Ground
-from roseau.load_flow.models.lines import Line, LineParameters, Switch
+from roseau.load_flow.models.lines import Line, LineParameters
 from roseau.load_flow.models.loads import (
     AbstractLoad,
     Control,
@@ -21,6 +21,7 @@ from roseau.load_flow.models.loads import (
 )
 from roseau.load_flow.models.potential_refs import PotentialRef
 from roseau.load_flow.models.sources import VoltageSource
+from roseau.load_flow.models.switches import Switch
 from roseau.load_flow.models.transformers import Transformer, TransformerParameters
 
 __all__ = [
@@ -32,9 +33,10 @@ __all__ = [
     # Buses
     "Bus",
     # Lines
-    "Switch",
     "Line",
     "LineParameters",
+    # Switches
+    "Switch",
     # Loads
     "AbstractLoad",
     "ImpedanceLoad",
