+++ roseau/load_flow/models/lines/__init__.py
@@ -1,4 +1,4 @@
-from roseau.load_flow.models.lines.lines import Line, Switch
+from roseau.load_flow.models.lines.lines import Line
 from roseau.load_flow.models.lines.parameters import LineParameters
 
-__all__ = ["Switch", "Line", "LineParameters"]
+__all__ = ["Line", "LineParameters"]
