+++ roseau/load_flow/models/lines/lines.py
@@ -8,116 +8,13 @@ from shapely.geometry.base import BaseGeometry
 from roseau.load_flow.exceptions import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
 from roseau.load_flow.models.branches import AbstractBranch
 from roseau.load_flow.models.buses import Bus
-from roseau.load_flow.models.core import Element
 from roseau.load_flow.models.grounds import Ground
 from roseau.load_flow.models.lines.parameters import LineParameters
-from roseau.load_flow.models.sources import VoltageSource
 from roseau.load_flow.typing import ComplexArray, Id, JsonDict
 from roseau.load_flow.units import Q_, ureg_wraps
-from roseau.load_flow_engine.cy_engine import CyShuntLine, CySimplifiedLine, CySwitch
+from roseau.load_flow_engine.cy_engine import CyShuntLine, CySimplifiedLine
 
 logger = logging.getLogger(__name__)
-
-
-class Switch(AbstractBranch):
-    """A general purpose switch branch."""
-
-    type: Final = "switch"
-
-    allowed_phases: Final = frozenset(Bus.allowed_phases | {"a", "b", "c", "n"})
-    """The allowed phases for a switch are:
-
-    - P-P-P or P-P-P-N: ``"abc"``, ``"abcn"``
-    - P-P or P-P-N: ``"ab"``, ``"bc"``, ``"ca"``, ``"abn"``, ``"bcn"``, ``"can"``
-    - P or P-N: ``"a"``, ``"b"``, ``"c"``, ``"an"``, ``"bn"``, ``"cn"``
-    - N: ``"n"``
-    """
-
-    def __init__(
-        self, id: Id, bus1: Bus, bus2: Bus, *, phases: str | None = None, geometry: BaseGeometry | None = None
-    ) -> None:
-        """Switch constructor.
-
-        Args:
-            id:
-                A unique ID of the switch in the network branches.
-
-            bus1:
-                Bus to connect to the switch.
-
-            bus2:
-                Bus to connect to the switch.
-
-            phases:
-                The phases of the switch. A string like ``"abc"`` or ``"an"`` etc. The order of the
-                phases is important. For a full list of supported phases, see the class attribute
-                :attr:`allowed_phases`. All phases of the switch must be present in the phases of
-                both connected buses. By default, the phases common to both buses are used.
-
-            geometry:
-                The geometry of the switch.
-        """
-        if phases is None:
-            phases = "".join(p for p in bus1.phases if p in bus2.phases)  # can't use set because order is important
-            phases = phases.replace("ac", "ca")
-        else:
-            # Also check they are in the intersection of buses phases
-            self._check_phases(id, phases=phases)
-            buses_phases = set(bus1.phases) & set(bus2.phases)
-            phases_not_in_buses = set(phases) - buses_phases
-            if phases_not_in_buses:
-                msg = (
-                    f"Phases {sorted(phases_not_in_buses)} of switch {id!r} are not in the common phases "
-                    f"{sorted(buses_phases)} of buses {bus1.id!r} and {bus2.id!r}."
-                )
-                logger.error(msg)
-                raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_PHASE)
-
-        super().__init__(id=id, phases1=phases, phases2=phases, bus1=bus1, bus2=bus2, geometry=geometry)
-        self._check_elements()
-        self._check_loop()
-        self._cy_element = CySwitch(self._n1)
-        self._cy_connect()
-
-    @property
-    def phases(self) -> str:
-        """The phases of the switch. This is an alias for :attr:`phases1` and :attr:`phases2`."""
-        return self._phases1
-
-    def _check_loop(self) -> None:
-        """Check that there are no switch loop, raise an exception if it is the case"""
-        visited_1: set[Element] = set()
-        elements: list[Element] = [self.bus1]
-        while elements:
-            element = elements.pop(-1)
-            visited_1.add(element)
-            for e in element._connected_elements:
-                if e not in visited_1 and (isinstance(e, (Bus, Switch))) and e != self:
-                    elements.append(e)
-        visited_2: set[Element] = set()
-        elements = [self.bus2]
-        while elements:
-            element = elements.pop(-1)
-            visited_2.add(element)
-            for e in element._connected_elements:
-                if e not in visited_2 and (isinstance(e, (Bus, Switch))) and e != self:
-                    elements.append(e)
-        if visited_1.intersection(visited_2):
-            msg = f"There is a loop of switch involving the switch {self.id!r}. It is not allowed."
-            logger.error(msg)
-            raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.SWITCHES_LOOP)
-
-    def _check_elements(self) -> None:
-        """Check that we can connect both elements."""
-        if any(isinstance(e, VoltageSource) for e in self.bus1._connected_elements) and any(
-            isinstance(e, VoltageSource) for e in self.bus2._connected_elements
-        ):
-            msg = (
-                f"The buses {self.bus1.id!r} and {self.bus2.id!r} both have a voltage source and "
-                f"are connected with the switch {self.id!r}. It is not allowed."
-            )
-            logger.error(msg)
-            raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.BAD_VOLTAGES_SOURCES_CONNECTION)
 
 
 class Line(AbstractBranch):
